+++ lib/features/item/detail/screen/item_detail_screen.dart
@@ -1,112 +1,39 @@
-import 'dart:async';
-
 import 'package:bidbird/core/managers/supabase_manager.dart';
-import 'package:bidbird/core/utils/ui_set/border_radius.dart';
-import 'package:bidbird/core/utils/ui_set/colors.dart';
-import 'package:bidbird/features/item/detail/data/datasource/item_detail_datasource.dart';
 import 'package:bidbird/features/item/detail/model/item_detail_entity.dart';
-import 'package:bidbird/features/item/price_Input/screen/price_input_screen.dart';
-import 'package:bidbird/features/item/price_Input/viewmodel/price_input_viewmodel.dart';
-import 'package:bidbird/features/item/buy_now_input/screen/buy_now_input_bottom_sheet.dart';
+import 'package:bidbird/features/item/detail/viewmodel/item_detail_viewmodel.dart';
 import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
-import 'package:supabase_flutter/supabase_flutter.dart';
 
 import '../../../report/ui/report_screen.dart';
+import 'widgets/item_image_section.dart';
+import 'widgets/item_main_info_section.dart';
+import 'widgets/item_description_section.dart';
+import 'widgets/item_bottom_action_bar.dart';
 
-class ItemDetailScreen extends StatefulWidget {
+class ItemDetailScreen extends StatelessWidget {
   const ItemDetailScreen({super.key, required this.itemId});
 
   final String itemId;
 
   @override
-  State<ItemDetailScreen> createState() => _ItemDetailScreenState();
+  Widget build(BuildContext context) {
+    return ChangeNotifierProvider<ItemDetailViewModel>(
+      create: (_) => ItemDetailViewModel(itemId: itemId)
+        ..loadItemDetail()
+        ..setupRealtimeSubscription(),
+      child: const _ItemDetailScaffold(),
+    );
+  }
 }
 
-class _ItemDetailScreenState extends State<ItemDetailScreen> {
-  late final SupabaseClient _supabase;
-  RealtimeChannel? _bidStatusChannel;
-  RealtimeChannel? _itemsChannel;
-  RealtimeChannel? _bidLogChannel;
-
-  @override
-  void initState() {
-    super.initState();
-    _supabase = SupabaseManager.shared.supabase;
-    _setupRealtimeSubscription();
-  }
-
-  @override
-  void dispose() {
-    if (_bidStatusChannel != null) _supabase.removeChannel(_bidStatusChannel!);
-    if (_itemsChannel != null) _supabase.removeChannel(_itemsChannel!);
-    if (_bidLogChannel != null) _supabase.removeChannel(_bidLogChannel!);
-    super.dispose();
-  }
-
-  void _setupRealtimeSubscription() {
-    // bid_status 테이블 실시간 구독
-    _bidStatusChannel = _supabase.channel('bid_status_${widget.itemId}');
-    _bidStatusChannel!
-        .onPostgresChanges(
-          event: PostgresChangeEvent.all,
-          schema: 'public',
-          table: 'bid_status',
-          filter: PostgresChangeFilter(
-            type: PostgresChangeFilterType.eq,
-            column: 'item_id',
-            value: widget.itemId,
-          ),
-          callback: (payload) {
-            if (mounted) setState(() {});
-          },
-        )
-        .subscribe();
-
-    // items 테이블 실시간 구독 (현재가 변경 감지)
-    _itemsChannel = _supabase.channel('items_${widget.itemId}');
-    _itemsChannel!
-        .onPostgresChanges(
-          event: PostgresChangeEvent.all,
-          schema: 'public',
-          table: 'items',
-          filter: PostgresChangeFilter(
-            type: PostgresChangeFilterType.eq,
-            column: 'id',
-            value: widget.itemId,
-          ),
-          callback: (payload) {
-            if (mounted) setState(() {});
-          },
-        )
-        .subscribe();
-
-    // bid_log 테이블 실시간 구독 (참여 입찰 수 변경 감지)
-    _bidLogChannel = _supabase.channel('bid_log_${widget.itemId}');
-    _bidLogChannel!
-        .onPostgresChanges(
-          event: PostgresChangeEvent.insert, // 입찰은 insert만 발생
-          schema: 'public',
-          table: 'bid_log',
-          filter: PostgresChangeFilter(
-            type: PostgresChangeFilterType.eq,
-            column: 'item_id',
-            value: widget.itemId,
-          ),
-          callback: (payload) {
-            if (mounted) setState(() {});
-          },
-        )
-        .subscribe();
-  }
+class _ItemDetailScaffold extends StatelessWidget {
+  const _ItemDetailScaffold();
 
   @override
   Widget build(BuildContext context) {
-    return FutureBuilder<ItemDetail?>(
-      future: _loadItemDetail(widget.itemId),
-      builder: (context, snapshot) {
-        if (snapshot.connectionState == ConnectionState.waiting) {
+    return Consumer<ItemDetailViewModel>(
+      builder: (context, vm, _) {
+        if (vm.isLoading) {
           return const Scaffold(
             body: SafeArea(
               child: Center(child: CircularProgressIndicator()),
@@ -114,17 +41,20 @@ class _ItemDetailScreenState extends State<ItemDetailScreen> {
           );
         }
 
-        if (snapshot.hasError || !snapshot.hasData || snapshot.data == null) {
+        if (vm.error != null || vm.itemDetail == null) {
           return const Scaffold(
             body: SafeArea(
               child: Center(
-                child: Text('매물 정보를 불러올 수 없습니다.', style: TextStyle(fontSize: 14)),
+                child: Text(
+                  '매물 정보를 불러올 수 없습니다.',
+                  style: TextStyle(fontSize: 14),
+                ),
               ),
             ),
           );
         }
 
-        final ItemDetail item = snapshot.data!;
+        final ItemDetail item = vm.itemDetail!;
 
         // 현재 로그인 유저와 판매자 비교해서 내 매물 여부 판단
         final supabase = SupabaseManager.shared.supabase;
@@ -176,16 +106,16 @@ class _ItemDetailScreenState extends State<ItemDetailScreen> {
                     child: Column(
                       crossAxisAlignment: CrossAxisAlignment.stretch,
                       children: [
-                        _ItemImageSection(item: item),
+                        ItemImageSection(item: item),
                         const SizedBox(height: 8),
-                        _ItemMainInfoSection(item: item, isMyItem: isMyItem),
+                        ItemMainInfoSection(item: item, isMyItem: isMyItem),
                         const SizedBox(height: 0),
-                        _ItemDescriptionSection(item: item),
+                        ItemDescriptionSection(item: item),
                       ],
                     ),
                   ),
                 ),
-                _BottomActionBar(item: item, isMyItem: isMyItem),
+                ItemBottomActionBar(item: item, isMyItem: isMyItem),
               ],
             ),
           ),
@@ -193,1028 +123,4 @@ class _ItemDetailScreenState extends State<ItemDetailScreen> {
       },
     );
   }
-}
-
-String _formatRelativeTime(String? isoString) {
-  if (isoString == null || isoString.isEmpty) return '';
-  DateTime? time;
-  try {
-    time = DateTime.tryParse(isoString);
-  } catch (_) {
-    time = null;
-  }
-  if (time == null) return '';
-
-  final now = DateTime.now();
-  final diff = now.difference(time);
-
-  if (diff.inSeconds < 60) {
-    return '방금 전';
-  } else if (diff.inMinutes < 60) {
-    return '${diff.inMinutes}분 전';
-  } else if (diff.inHours < 24) {
-    return '${diff.inHours}시간 전';
-  } else {
-    final days = diff.inDays;
-    return '${days}일 전';
-  }
-}
-
-Future<ItemDetail?> _loadItemDetail(String itemId) async {
-  final supabase = SupabaseManager.shared.supabase;
-
-  final List<dynamic> result = await supabase
-      .from('items')
-      .select()
-      .eq('id', itemId)
-      .limit(1);
-
-  if (result.isEmpty) {
-    return null;
-  }
-
-  final row = result.first;
-
-  // created_at + auction_duration_hours 로 종료 시각 계산
-  final createdAtRaw = row['created_at']?.toString();
-  final createdAt = createdAtRaw != null
-      ? DateTime.tryParse(createdAtRaw) ?? DateTime.now()
-      : DateTime.now();
-
-  final durationHours = (row['auction_duration_hours'] as int?) ?? 24;
-  final finishTime = createdAt.add(Duration(hours: durationHours));
-
-  final String sellerId = row['seller_id']?.toString() ?? '';
-  String sellerTitle = '';
-  double sellerRating = 0.0;
-  int sellerReviewCount = 0;
-
-  if (sellerId.isNotEmpty) {
-    try {
-      // users 테이블에서 모든 필드 조회
-      final userResponse = await supabase
-          .from('users')
-          .select()
-          .eq('id', sellerId)
-          .single();
-
-      // 닉네임 필드 확인
-      String? nickname = userResponse['nick_name']?.toString().trim();
-
-      if (nickname != null && nickname.isNotEmpty) {
-        sellerTitle = nickname;
-      } else {
-        sellerTitle = '미지정 사용자';
-      }
-
-      // items 테이블 업데이트
-      try {
-        await supabase
-            .from('items')
-            .update({'seller_name': sellerTitle})
-            .eq('id', itemId);
-      } catch (e) {
-        debugPrint('Failed to update seller_name for itemId=$itemId: $e');
-      }
-    } catch (e) {
-      // users 테이블 조회 실패 시 items 테이블의 seller_name 사용
-      sellerTitle = row['seller_name']?.toString() ?? '미지정 사용자';
-    }
-  } else {
-    sellerTitle = '알 수 없는 판매자';
-  }
-
-  // sellerId 기준으로 user_review에서 평점/리뷰 수 계산
-  if (sellerId.isNotEmpty) {
-    try {
-      final reviews = await supabase
-          .from('user_review')
-          .select('rating')
-          .eq('to_user_id', sellerId)
-          .not('rating', 'is', null);
-
-      if (reviews is List) {
-        double total = 0;
-        int count = 0;
-        for (final row in reviews) {
-          final value = (row['rating'] as num?)?.toDouble();
-          if (value != null) {
-            total += value;
-            count += 1;
-          }
-        }
-        sellerReviewCount = count;
-        sellerRating = count > 0 ? total / count : 0.0;
-      }
-    } catch (e) {
-      debugPrint('Failed to load seller rating in _loadItemDetail: $e');
-      sellerRating = 0.0;
-      sellerReviewCount = 0;
-    }
-  }
-
-  // item_images 테이블에서 이미지 URL 가져오기 (썸네일 제외)
-  final List<String> images = [];
-
-  try {
-    final imageRows = await supabase
-        .from('item_images')
-        .select('image_url')
-        .eq('item_id', itemId)
-        .order('sort_order', ascending: true);
-
-    for (final Map<String, dynamic> row in imageRows) {
-      final imageUrl = row['image_url']?.toString();
-      if (imageUrl != null && imageUrl.isNotEmpty) {
-        images.add(imageUrl);
-      }
-    }
-  } catch (e) {
-    debugPrint('Failed to load item images for itemId=$itemId: $e');
-  }
-
-  // bid_log 테이블에서 참여 입찰 수 조회
-  int biddingCount = 0;
-  try {
-    final countResponse = await supabase
-        .from('bid_log')
-        .select('id')
-        .eq('item_id', itemId)
-        .count(CountOption.exact);
-
-    biddingCount = countResponse.count;
-  } catch (e) {
-    // count 쿼리가 실패하면 기존 방식(items 테이블의 bidding_count) 사용하거나 0으로 설정
-    biddingCount = (row['bidding_count'] as int?) ?? 0;
-  }
-
-  // 입찰 최소 호가 규칙 적용
-  final currentPrice = (row['current_price'] as int?) ?? 0;
-  int minBidStep;
-
-  if (currentPrice <= 100000) {
-    // 100,000원 이하일 경우 최소 호가는 1,000원으로 고정
-    minBidStep = 1000;
-  } else {
-    // 100,001원 이상일 경우 마지막 두 자리 제거
-    final priceStr = currentPrice.toString();
-    if (priceStr.length >= 3) {
-      minBidStep = int.parse(priceStr.substring(0, priceStr.length - 2));
-    } else {
-      minBidStep = 1000; // 최소값 보장
-    }
-  }
-
-
-  return ItemDetail(
-    itemId: row['id']?.toString() ?? itemId,
-    sellerId: sellerId,
-    itemTitle: row['title']?.toString() ?? '',
-    itemImages: images,
-    finishTime: finishTime,
-    sellerTitle: sellerTitle,
-    buyNowPrice: (row['buy_now_price'] as int?) ?? 0,
-    biddingCount: biddingCount,
-    itemContent: row['description']?.toString() ?? '',
-    currentPrice: currentPrice,
-    bidPrice: minBidStep, // 계산된 최소 호가 사용
-    sellerRating: sellerRating,
-    sellerReviewCount: sellerReviewCount,
-    statusCode: (row['status_code'] as int?) ?? 0,
-  );
-}
-
-class _ItemImageSection extends StatefulWidget {
-  const _ItemImageSection({required this.item});
-
-  final ItemDetail item;
-
-  @override
-  State<_ItemImageSection> createState() => _ItemImageSectionState();
-}
-
-class _ItemImageSectionState extends State<_ItemImageSection> {
-  int _currentPage = 0;
-  final PageController _pageController = PageController();
-
-  @override
-  void dispose() {
-    _pageController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final hasImages = widget.item.itemImages.isNotEmpty;
-    final images = hasImages ? widget.item.itemImages : <String>[];
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        AspectRatio(
-          aspectRatio: 1,
-          child: Stack(
-            children: [
-              if (hasImages && images.isNotEmpty)
-                PageView.builder(
-                  controller: _pageController,
-                  onPageChanged: (index) {
-                    setState(() {
-                      _currentPage = index;
-                    });
-                  },
-                  itemCount: images.length,
-                  itemBuilder: (context, index) {
-                    return Container(
-                      width: double.infinity,
-                      color: ImageBackgroundColor,
-                      child: Image.network(
-                        images[index],
-                        width: double.infinity,
-                        fit: BoxFit.cover,
-                        errorBuilder: (context, error, stackTrace) {
-                          return const Center(
-                            child: Text(
-                              '상품 사진',
-                              style: TextStyle(color: Colors.grey),
-                            ),
-                          );
-                        },
-                      ),
-                    );
-                  },
-                )
-              else
-                Container(
-                  width: double.infinity,
-                  color: ImageBackgroundColor,
-                  child: const Center(
-                    child: Text(
-                      '상품 사진',
-                      style: TextStyle(color: iconColor),
-                    ),
-                  ),
-                ),
-              Positioned(
-                top: 16,
-                left: 16,
-                child: Container(
-                  padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                  decoration: BoxDecoration(
-                    color: RedColor,
-                    borderRadius: BorderRadius.circular(4),
-                  ),
-                  child: Text(
-                    '${_formatRemainingTime(widget.item.finishTime)} 남음',
-                    style: const TextStyle(
-                      color: Colors.white,
-                      fontSize: 12,
-                      fontWeight: FontWeight.w600,
-                    ),
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-        if (hasImages && images.length > 1) ...[
-          const SizedBox(height: 8),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: List.generate(
-              images.length,
-              (index) => _buildDot(isActive: index == _currentPage),
-            ),
-          ),
-        ],
-      ],
-    );
-  }
-
-  Widget _buildDot({required bool isActive}) {
-    return Container(
-      width: 6,
-      height: 6,
-      margin: const EdgeInsets.symmetric(horizontal: 3),
-      decoration: BoxDecoration(
-        shape: BoxShape.circle,
-        // 현재 사진: 안쪽은 검은색, 나머지는 투명
-        color: isActive ? Colors.black : Colors.transparent,
-        border: Border.all(
-          color: isActive ? Colors.black : const Color(0xFFBDBDBD),
-          width: isActive ? 1.5 : 1,
-        ),
-      ),
-    );
-  }
-}
-
-class _ItemMainInfoSection extends StatelessWidget {
-  const _ItemMainInfoSection({required this.item, required this.isMyItem});
-
-  final ItemDetail item;
-  final bool isMyItem;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: const EdgeInsets.fromLTRB(16, 16, 16, 16),
-      decoration: const BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.only(
-          topLeft: Radius.circular(defaultRadius),
-          topRight: Radius.circular(defaultRadius),
-        ),
-      ),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Row(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              Expanded(
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      item.itemTitle,
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                      style: const TextStyle(
-                        fontSize: 18,
-                        fontWeight: FontWeight.w700,
-                      ),
-                    ),
-                    const SizedBox(height: 4),
-                    Text(
-                      item.buyNowPrice > 0
-                          ? '즉시 구매가 ${_formatPrice(item.buyNowPrice)}원'
-                          : '즉시 구매 없음',
-                      style: TextStyle(
-                        fontSize: 13,
-                        fontWeight: FontWeight.w600,
-                        color: item.buyNowPrice > 0
-                            ? blueColor
-                            : iconColor,
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              const SizedBox(width: 8),
-              if (!isMyItem)
-                TextButton.icon(
-                  onPressed: () {
-                    // TODO: 판매자 연락 기능 연동 (채팅 등)
-                  },
-                  style: TextButton.styleFrom(
-                    padding: const EdgeInsets.symmetric(horizontal: 4),
-                    minimumSize: const Size(0, 0),
-                    tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                  ),
-                  icon: const Icon(
-                    Icons.chat_bubble_outline,
-                    size: 18,
-                    color: blueColor,
-                  ),
-                  label: const Text(
-                    '판매자 연락',
-                    style: TextStyle(
-                      fontSize: 12,
-                      fontWeight: FontWeight.w500,
-                      color: blueColor,
-                    ),
-                  ),
-                ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          Container(
-            padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 14),
-            decoration: BoxDecoration(
-              color: BorderColor.withValues(alpha: 0.3),
-              borderRadius: BorderRadius.circular(defaultRadius),
-              boxShadow: const [],
-            ),
-            child: Row(
-              children: [
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      const Text(
-                        '현재 입찰가',
-                        style: TextStyle(
-                          fontSize: 12,
-                          color: iconColor,
-                        ),
-                      ),
-                      const SizedBox(height: 4),
-                      Text(
-                        '${_formatPrice(item.currentPrice)}원',
-                        style: const TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.w700,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-                Container(
-                  width: 1,
-                  height: 36,
-                  color: BorderColor.withValues(alpha: 0.5),
-                ),
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.end,
-                    children: [
-                      const Text(
-                        '참여 입찰',
-                        style: TextStyle(
-                          fontSize: 12,
-                          color: iconColor,
-                        ),
-                      ),
-                      const SizedBox(height: 4),
-                      Text(
-                        '${item.biddingCount}건',
-                        style: const TextStyle(
-                          fontSize: 20,
-                          fontWeight: FontWeight.w700,
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _ItemDescriptionSection extends StatelessWidget {
-  const _ItemDescriptionSection({required this.item});
-
-  final ItemDetail item;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 0, 16, 16),
-          child: Container(
-            padding: const EdgeInsets.fromLTRB(16, 12, 16, 16),
-            decoration: BoxDecoration(
-              color: BorderColor.withValues(alpha: 0.3),
-              borderRadius: BorderRadius.circular(defaultRadius),
-              boxShadow: const [],
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                // 판매자 프로필 영역
-                Row(
-                  children: [
-                    const CircleAvatar(
-                      radius: 20,
-                      backgroundColor: yellowColor,
-                      child: Icon(Icons.person, color: BackgroundColor),
-                    ),
-                    const SizedBox(width: 12),
-                    Expanded(
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          Text(
-                            item.sellerTitle,
-                            style: const TextStyle(
-                              fontSize: 14,
-                              fontWeight: FontWeight.w600,
-                            ),
-                          ),
-                          const SizedBox(height: 4),
-                          Row(
-                            children: [
-                              const Icon(
-                                Icons.star,
-                                size: 14,
-                                color: yellowColor,
-                              ),
-                              const SizedBox(width: 4),
-                              Text(
-                                '${item.sellerRating.toStringAsFixed(1)} (${item.sellerReviewCount})',
-                                style: const TextStyle(
-                                  fontSize: 12,
-                                  color: iconColor,
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
-                    ),
-                    TextButton(
-                      onPressed: () {
-                        if (item.sellerId.isEmpty) return;
-                        context.push('/user/${item.sellerId}');
-                      },
-                      style: TextButton.styleFrom(
-                        padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 6),
-                        tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                        minimumSize: const Size(0, 0),
-                      ),
-                      child: Row(
-                        mainAxisSize: MainAxisSize.min,
-                        children: const [
-                          Text(
-                            '프로필 보기',
-                            style: TextStyle(
-                              fontSize: 12,
-                              fontWeight: FontWeight.w600,
-                              color: blueColor,
-                            ),
-                          ),
-                          SizedBox(width: 4),
-                          Icon(
-                            Icons.chevron_right,
-                            size: 16,
-                            color: blueColor,
-                          ),
-                        ],
-                      ),
-                    ),
-                  ],
-                ),
-                const SizedBox(height: 12),
-                const Divider(
-                  height: 1,
-                  thickness: 1,
-                  color: BorderColor,
-                ),
-                const SizedBox(height: 12),
-                // 상품 설명 영역
-                const Text(
-                  '상품 설명',
-                  style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
-                ),
-                const SizedBox(height: 8),
-                Text(
-                  item.itemContent,
-                  style: const TextStyle(fontSize: 13, height: 1.4),
-                ),
-              ],
-            ),
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.fromLTRB(16, 0, 16, 24),
-          child: Container(
-            width: double.infinity,
-            padding: const EdgeInsets.fromLTRB(16, 12, 16, 16),
-            decoration: BoxDecoration(
-              color: BorderColor.withValues(alpha: 0.3),
-              borderRadius: BorderRadius.circular(defaultRadius),
-              boxShadow: const [],
-            ),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                const Text(
-                  '현재 입찰 내역',
-                  style: TextStyle(fontSize: 14, fontWeight: FontWeight.w600),
-                ),
-                const SizedBox(height: 8),
-                FutureBuilder<List<Map<String, dynamic>>>(
-                  future: ItemDetailDatasource().fetchBidHistory(item.itemId),
-                  builder: (context, snapshot) {
-                    if (snapshot.connectionState == ConnectionState.waiting) {
-                      return const Center(
-                        child: SizedBox(
-                          height: 24,
-                          width: 24,
-                          child: CircularProgressIndicator(strokeWidth: 2),
-                        ),
-                      );
-                    }
-
-                    if (snapshot.hasError) {
-                      return const Text(
-                        '입찰 내역을 불러올 수 없습니다.',
-                        style: TextStyle(fontSize: 12, color: iconColor),
-                      );
-                    }
-
-                    final bids = snapshot.data ?? [];
-                    if (bids.isEmpty) {
-                      return const Text(
-                        '아직 입찰 내역이 없습니다.',
-                        style: TextStyle(fontSize: 12, color: iconColor),
-                      );
-                    }
-
-                    final limited = bids.take(10).toList();
-
-                    return ListView.separated(
-                      shrinkWrap: true,
-                      physics: const NeverScrollableScrollPhysics(),
-                      itemCount: limited.length,
-                      separatorBuilder: (_, __) => const SizedBox(height: 6),
-                      itemBuilder: (context, index) {
-                        final bid = limited[index];
-                        final price = bid['price']?.toString() ?? '';
-                        final createdAtRaw = bid['created_at']?.toString();
-                        final relative = _formatRelativeTime(createdAtRaw);
-
-                        return Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            Text(
-                              '${index + 1}. $price원',
-                              style: const TextStyle(
-                                fontSize: 13,
-                                fontWeight: FontWeight.w600,
-                              ),
-                            ),
-                            Text(
-                              relative,
-                              style: const TextStyle(
-                                fontSize: 11,
-                                color: iconColor,
-                              ),
-                            ),
-                          ],
-                        );
-                      },
-                    );
-                  },
-                ),
-              ],
-            ),
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class _BottomActionBar extends StatefulWidget {
-  const _BottomActionBar({required this.item, required this.isMyItem});
-
-  final ItemDetail item;
-  final bool isMyItem;
-
-  @override
-  State<_BottomActionBar> createState() => _BottomActionBarState();
-}
-
-class _BottomActionBarState extends State<_BottomActionBar> {
-  bool _isFavorite = false;
-  bool _isTopBidder = false;
-  int? _statusCode; // bid_status.int_code 기반 현재 상태 코드
-
-  @override
-  void initState() {
-    super.initState();
-    _statusCode = widget.item.statusCode;
-    _loadFavoriteState();
-    _checkTopBidder();
-  }
-
-  Future<void> _loadFavoriteState() async {
-    final supabase = SupabaseManager.shared.supabase;
-    final user = supabase.auth.currentUser;
-    if (user == null) return;
-
-    try {
-      final List<dynamic> rows = await supabase
-          .from('favorites')
-          .select('id')
-          .eq('item_id', widget.item.itemId)
-          .eq('user_id', user.id)
-          .limit(1);
-
-      if (!mounted) return;
-      setState(() {
-        _isFavorite = rows.isNotEmpty;
-      });
-    } catch (e) {
-      debugPrint(
-        'Failed to load favorite state for itemId=${widget.item.itemId}: $e',
-      );
-    }
-  }
-
-  Future<void> _checkTopBidder() async {
-    final supabase = SupabaseManager.shared.supabase;
-    final user = supabase.auth.currentUser;
-    if (user == null) return;
-
-    try {
-      final row = await supabase
-          .from('bid_status')
-          .select('current_highest_bidder, int_code')
-          .eq('item_id', widget.item.itemId)
-          .maybeSingle();
-
-      if (!mounted || row == null) return;
-
-      final String? currentHighest =
-          row['current_highest_bidder']?.toString();
-      final int? intCode = row['int_code'] as int?;
-
-      setState(() {
-        _isTopBidder = currentHighest != null && currentHighest == user.id;
-        if (intCode != null) {
-          _statusCode = intCode;
-        }
-      });
-    } catch (e) {
-      debugPrint(
-        'Failed to check top bidder for itemId=${widget.item.itemId}: $e',
-      );
-    }
-  }
-
-  Future<void> _toggleFavorite() async {
-    final supabase = SupabaseManager.shared.supabase;
-    final user = supabase.auth.currentUser;
-    if (user == null) return;
-
-    try {
-      if (_isFavorite) {
-        await supabase
-            .from('favorites')
-            .delete()
-            .eq('item_id', widget.item.itemId)
-            .eq('user_id', user.id);
-      } else {
-        await supabase.from('favorites').insert(<String, dynamic>{
-          'item_id': widget.item.itemId,
-          'user_id': user.id,
-        });
-      }
-
-      if (!mounted) return;
-      setState(() {
-        _isFavorite = !_isFavorite;
-      });
-    } catch (e) {
-      debugPrint(
-        'Failed to toggle favorite for itemId=${widget.item.itemId}: $e',
-      );
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // status_code 기준 즉시 구매 상태
-    final int statusCode = _statusCode ?? widget.item.statusCode;
-    final bool isBuyNowPending = statusCode == 1006; // 즉시 구매 대기
-    final bool isBuyNowCompleted = statusCode == 1007; // 즉시 구매 완료
-
-    final bool showBidButton =
-        !_isTopBidder && !isBuyNowPending && !isBuyNowCompleted;
-    final bool showBuyNowButton = widget.item.buyNowPrice > 0 &&
-        !isBuyNowPending && !isBuyNowCompleted;
-
-    return SafeArea(
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-        color: BackgroundColor,
-        child: Row(
-          children: [
-            Expanded(
-              child: Row(
-                children: [
-                  if (!widget.isMyItem) ...[
-                    SizedBox(
-                      width: 44,
-                      height: 44,
-                      child: IconButton(
-                        onPressed: _toggleFavorite,
-                        padding: EdgeInsets.zero,
-                        splashRadius: 24,
-                        icon: Icon(
-                          _isFavorite ? Icons.favorite : Icons.favorite_border,
-                          size: 24,
-                          color: RedColor,
-                        ),
-                      ),
-                    ),
-                    const SizedBox(width: 8),
-
-                    // 입찰하기 / 상태 텍스트
-                    Expanded(
-                      child: SizedBox(
-                        height: 44,
-                        child: _buildBidOrStatusContent(
-                          isBuyNowPending: isBuyNowPending,
-                          isBuyNowCompleted: isBuyNowCompleted,
-                          showBidButton: showBidButton,
-                        ),
-                      ),
-                    ),
-
-                    if (showBuyNowButton) ...[
-                      const SizedBox(width: 8),
-                      Expanded(
-                        child: SizedBox(
-                          height: 44,
-                          child: _buildBuyNowButton(),
-                        ),
-                      ),
-                    ],
-                  ],
-                ],
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildBidOrStatusContent({
-    required bool isBuyNowPending,
-    required bool isBuyNowCompleted,
-    required bool showBidButton,
-  }) {
-    // 1) 내가 최고 입찰자인 경우
-    if (_isTopBidder) {
-      return Container(
-        decoration: BoxDecoration(
-          color: BackgroundColor,
-          borderRadius: BorderRadius.circular(8.7),
-          border: Border.all(color: BorderColor),
-        ),
-        child: Center(
-          child: Text(
-            '최고 입찰자입니다',
-            style: TextStyle(
-              fontSize: 13,
-              fontWeight: FontWeight.w600,
-              color: TopBidderTextColor,
-            ),
-          ),
-        ),
-      );
-    }
-
-    // 2) 즉시 구매 진행 중
-    if (isBuyNowPending) {
-      return Container(
-        decoration: BoxDecoration(
-          color: BackgroundColor,
-          borderRadius: BorderRadius.circular(8.7),
-          border: Border.all(color: BorderColor),
-        ),
-        child: const Center(
-          child: Text(
-            '즉시 구매 중입니다',
-            style: TextStyle(
-              fontSize: 13,
-              fontWeight: FontWeight.w600,
-              color: TopBidderTextColor,
-            ),
-          ),
-        ),
-      );
-    }
-
-    // 3) 즉시 구매 완료
-    if (isBuyNowCompleted) {
-      return Container(
-        decoration: BoxDecoration(
-          color: BackgroundColor,
-          borderRadius: BorderRadius.circular(8.7),
-          border: Border.all(color: BorderColor),
-        ),
-        child: const Center(
-          child: Text(
-            '즉시 구매되었습니다',
-            style: TextStyle(
-              fontSize: 13,
-              fontWeight: FontWeight.w600,
-              color: TopBidderTextColor,
-            ),
-          ),
-        ),
-      );
-    }
-
-    // 4) 그 외에는 입찰하기 버튼 표시
-    if (showBidButton) {
-      return OutlinedButton(
-        onPressed: () {
-          showModalBottomSheet(
-            context: context,
-            isScrollControlled: true,
-            backgroundColor: Colors.white,
-            shape: const RoundedRectangleBorder(
-              borderRadius: BorderRadius.vertical(
-                top: Radius.circular(defaultRadius),
-              ),
-            ),
-            builder: (context) {
-              return ChangeNotifierProvider<PriceInputViewModel>(
-                create: (_) => PriceInputViewModel(),
-                child: BidBottomSheet(
-                  itemId: widget.item.itemId,
-                  currentPrice: widget.item.currentPrice,
-                  bidUnit: widget.item.bidPrice,
-                  buyNowPrice: widget.item.buyNowPrice,
-                ),
-              );
-            },
-          );
-        },
-        style: OutlinedButton.styleFrom(
-          side: BorderSide(color: blueColor),
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(8.7),
-          ),
-        ),
-        child: Text(
-          '입찰하기',
-          style: TextStyle(
-            fontSize: 13,
-            fontWeight: FontWeight.w600,
-            color: blueColor,
-          ),
-        ),
-      );
-    }
-
-    // 방어 코드: 이론상 여기 안 옴
-    return const SizedBox.shrink();
-  }
-
-  Widget _buildBuyNowButton() {
-    return ElevatedButton(
-      onPressed: () {
-        showModalBottomSheet(
-          context: context,
-          isScrollControlled: true,
-          backgroundColor: Colors.white,
-          shape: const RoundedRectangleBorder(
-            borderRadius: BorderRadius.vertical(
-              top: Radius.circular(defaultRadius),
-            ),
-          ),
-          builder: (context) {
-            return ChangeNotifierProvider<PriceInputViewModel>(
-              create: (_) => PriceInputViewModel(),
-              child: BuyNowInputBottomSheet(
-                itemId: widget.item.itemId,
-                buyNowPrice: widget.item.buyNowPrice,
-              ),
-            );
-          },
-        );
-      },
-      style: ElevatedButton.styleFrom(
-        backgroundColor: blueColor,
-        shape: RoundedRectangleBorder(
-          borderRadius: BorderRadius.circular(8.7),
-        ),
-      ),
-      child: const Text(
-        '즉시 구매하기',
-        style: TextStyle(
-          fontSize: 13,
-          fontWeight: FontWeight.w600,
-          color: Colors.white,
-        ),
-      ),
-    );
-  }
-}
-
-String _formatRemainingTime(DateTime finishTime) {
-  final diff = finishTime.difference(DateTime.now());
-  if (diff.isNegative) {
-    return '00:00';
-  }
-  final hours = diff.inHours;
-  final minutes = diff.inMinutes % 60;
-  return '${hours.toString().padLeft(2, '0')}:${minutes.toString().padLeft(2, '0')}';
-}
-
-String _formatPrice(int price) {
-  // 콤마 포맷팅
-  final buffer = StringBuffer();
-  final text = price.toString();
-  for (int i = 0; i < text.length; i++) {
-    final reverseIndex = text.length - i;
-    buffer.write(text[i]);
-    if (reverseIndex > 1 && reverseIndex % 3 == 1 && i != text.length - 1) {
-      buffer.write(',');
-    }
-  }
-  return buffer.toString();
 }
