+++ lib/features/item/detail/screen/item_detail_utils.dart
@@ -1,0 +1,47 @@
+String formatRemainingTime(DateTime finishTime) {
+  final diff = finishTime.difference(DateTime.now());
+  if (diff.isNegative) {
+    return '00:00';
+  }
+  final hours = diff.inHours;
+  final minutes = diff.inMinutes % 60;
+  return '${hours.toString().padLeft(2, '0')}:${minutes.toString().padLeft(2, '0')}';
+}
+
+String formatPrice(int price) {
+  final buffer = StringBuffer();
+  final text = price.toString();
+  for (int i = 0; i < text.length; i++) {
+    final reverseIndex = text.length - i;
+    buffer.write(text[i]);
+    if (reverseIndex > 1 && reverseIndex % 3 == 1 && i != text.length - 1) {
+      buffer.write(',');
+    }
+  }
+  return buffer.toString();
+}
+
+String formatRelativeTime(String? isoString) {
+  if (isoString == null || isoString.isEmpty) return '';
+  DateTime? time;
+  try {
+    time = DateTime.tryParse(isoString);
+  } catch (_) {
+    time = null;
+  }
+  if (time == null) return '';
+
+  final now = DateTime.now();
+  final diff = now.difference(time);
+
+  if (diff.inSeconds < 60) {
+    return '방금 전';
+  } else if (diff.inMinutes < 60) {
+    return '${diff.inMinutes}분 전';
+  } else if (diff.inHours < 24) {
+    return '${diff.inHours}시간 전';
+  } else {
+    final days = diff.inDays;
+    return '${days}일 전';
+  }
+}
